Viola-Jones-Program/requirements.txt
numpy==1.21.6
opencv-python==4.5.5.64
Pillow==9.1.0
